Add org.gradle.alternateDistDir property for gradle-wrapper
When the Gradle Wrapper fails to find a distribution by trying to
fetch the distributionUrl, and if the system property
org.gradle.alternateDistDir is set, then the wrapper will attempt
to fetch the distribution from the alternate local directory.

This allows for building in an environment that does not have
access to the original URL in distributionUrl, without requiring
changing gradle-wrapper.properties.